share interrupt state (#1148)
